test: split monolithic health-check.test.mjs into logical groups
Closes #296

Break the 1313-line health-check.test.mjs (111 tests across 33
describe blocks) into six topic-focused files:

- cleaner-domains.test.mjs: per-domain URL cleaning + regressions
- config-integrity.test.mjs: domain-rules.json + manifest.json
- firefox-mv2.test.mjs: MV2 compat guards, polyfill, manifest
- i18n-integrity.test.mjs: i18n keys + HTML data-i18n sync
- onboarding.test.mjs: consent gate + onboarding dedup
- misc-regression.test.mjs: storage flush + param categories smoke

Pure refactor — each test body copied verbatim, imports
tree-shaken per file, zero logic changes. All 964 tests pass.

Completes Phase 6.4 of the code-quality plan.
